// projects/PieOut-frontend/src/types/ModalStateProps.ts
// Define a modal state type - each modal has a boolean open/closed state
export type ModalTypes = {
  wallet: boolean
  profile: boolean
  profileBlurb: boolean
  unregister: boolean
  game: boolean
  gameBlurb: boolean
  honors: boolean
  honorsBlurb: boolean
  activePlayers: boolean
  leaderboard: boolean
}